Rename “plugins” to “configure”
