+++ src/configure/index.spec.js
