+++ src/configure/karma/index.js
