+++ src/configure/path/index.js
