+++ src/configure/webpack/babel-plugin.js
