+++ src/configure/webpack/base-plugin.js
