+++ src/configure/webpack/css-modules-plugin.js
