+++ src/configure/webpack/eslint-plugin.js
