+++ src/configure/webpack/json-plugin.js
