+++ src/configure/webpack/media-plugin.js
