+++ src/configure/webpack/pages-plugin.js
