+++ src/index.js
@@ -1,29 +1,29 @@
 import * as actions from './actions'
-import plugins from './plugins'
+import configure from './configure'
 
 export default {
   build (options) {
-    const env = plugins({ ...options, buildTarget: 'build' })
+    const env = configure({ ...options, buildTarget: 'build' })
     actions.buildAndDistribute(env)
   },
 
   dist (options) {
-    const env = plugins({ ...options, buildTarget: 'dist' })
+    const env = configure({ ...options, buildTarget: 'dist' })
     actions.buildAndDistribute(env)
   },
 
   develop (options) {
-    const env = plugins({ ...options, buildTarget: 'develop' })
+    const env = configure({ ...options, buildTarget: 'develop' })
     actions.startDevelop(env)
   },
 
   test (options) {
-    const env = plugins({ ...options, buildTarget: 'test' })
+    const env = configure({ ...options, buildTarget: 'test' })
     actions.runTest(env)
   },
 
   install (options) {
-    const env = plugins(options)
+    const env = configure(options)
     actions.install(env)
   }
 }
